+++ vue/components/ui/molecules/index.js
@@ -22,6 +22,7 @@ import { ImageList } from "./image-list/image-list.vue";
 import { InputColor } from "./input-color/input-color.vue";
 import { InputCurrency } from "./input-currency/input-currency.vue";
 import { InputDate } from "./input-date/input-date.vue";
+import { InputFile } from "./input-file/input-file.vue";
 import { InputImage } from "./input-image/input-image.vue";
 import { InputSlider } from "./input-slider/input-slider.vue";
 import { InputSymbol } from "./input-symbol/input-symbol.vue";
@@ -98,6 +99,7 @@ const install = Vue => {
     Vue.component("tooltip", Tooltip);
     Vue.component("transfer-list", TransferList);
     Vue.component("upload-area", UploadArea);
+    Vue.component("input-file", InputFile);
 };
 
 export {
@@ -124,6 +126,7 @@ export {
     InputColor,
     InputCurrency,
     InputDate,
+    InputFile,
     InputImage,
     InputSlider,
     InputSymbol,